Change mergeOptions() to merge instead of replacing by key.
In my use case, we've extended FormBuilder\Fields\SelectType and we're supplying values for $options['rules'] and $options['rules_append'] in in getDefaults( ) and the consructor respectively, but these are being removed if there are values set in $form->add(... 'rules' => [...]).

This is because on FormField line 273, we use this helper to merge options from the class and those from $form->add(...). array_replace_recursive(), however, replaces based on key value. And key value is just basic integers. Instead, we need to array_merge_recursive().